backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rahul Samala <samalarahul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^4.2.8"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
langchain = "^0.3.19"
langchain-community = "^0.3.18"
ollama = "^0.4.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
